+++ Functions/SOI3Dviz.py
