+++ Functions/soi.py
