+++ Functions/viewpoints.py
